project/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pyspellchecker>=0.7.2
nltk>=3.8.1
numpy>=1.24.0
